# docs/openapi.yaml
openapi: 3.1.0
info:
  title: SERA Core API
  description: REST API for SERA Core service managing agents, circles, memory, skills, and sandbox environments.
  version: 1.0.0
servers:
  - url: http://localhost:3001/api
    description: Local Development
  - url: http://sera-core:3001/api
    description: Docker Deployment

# Empty security array applies globally so security-defined passes,
# as SERA Core doesn't appear to mandate global api key checks natively at this API level
security: []
paths:
...
  /lsp/definition:
    post:
      tags:
        - LSP
      summary: Get definition
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                filePath:
                  type: string
                line:
                  type: integer
                character:
                  type: integer
              required:
                - filePath
                - line
                - character
      responses:
        '200':
          description: Definition
          content:
            application/json:
              schema:
                type: object
                properties:
                  definition:
                    type: object
  /lsp/references:
    post:
      tags:
        - LSP
      summary: Get references
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                filePath:
                  type: string
                line:
                  type: integer
                character:
                  type: integer
              required:
                - filePath
                - line
                - character
      responses:
        '200':
          description: References
          content:
            application/json:
              schema:
                type: object
                properties:
                  references:
                    type: array
                    items:
                      type: object
  /lsp/symbols:
    post:
      tags:
        - LSP
      summary: Get document symbols
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                filePath:
                  type: string
              required:
                - filePath
      responses:
        '200':
          description: Symbols
          content:
            application/json:
              schema:
                type: object
                properties:
                  symbols:
                    type: array
                    items:
                      type: object
...